refactor: 클래스화
def 묶어서 하나의 클래스로 리팩토링
